+++ src/main/java/com/example/shop/repository/ItemRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.querydsl.QuerydslPredicateExecutor;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
@@ -12,5 +13,5 @@ public interface ItemRepository extends JpaRepository<Item, Long>, ItemRepositor
 
     List<Item> findByItemName(String itemName);
 
-    List<Item> findByPriceLessThan(Integer price);
+    List<Item> findByPriceLessThan(BigDecimal price);
 }
